Move `rray_increase_dims()` and `r_identical()` to their respective files
