+++ inst/include/api.h
@@ -3,6 +3,16 @@
 
 #include <Rcpp.h>
 
+// -----------------------------------------------------------------------------
+// Dimensions / dimensionality
+
 Rcpp::IntegerVector rray_dim(Rcpp::RObject x);
+
+Rcpp::IntegerVector rray_increase_dims(const Rcpp::IntegerVector& dim, const int& dims);
+
+// -----------------------------------------------------------------------------
+// Re-exposed R API
+
+bool r_identical(SEXP x, SEXP y);
 
 #endif
